hww: move actual api processing to new file
Separate the files a bitm so hww.c is doing arbitration (HWW_REQ_*
level) and hww_api.c is doing the actual api processing.
